Switch to kwargs and lower case the epsg spec
Great notebook! A couple things: pyproj lets you use keyword arguments in the `Proj()` constructor, which is better Python. Also, you need to lowercase the epsg identifiers for this to work on case-sensitive filesystems like Linux. On OS X, `EPSG:4326' tells PROJ.4 to look for the 4326 record in a file named "EPSG", and since OS X is case-insensitive, it finds and opens the actual file named "epsg".